==> code/Notebook-2-Data-Cleaning.ipynb <==
# Project 3: Web APIs & NLP
# Notebook 2: Data Cleaning

https://github.com/pushshift/api<br>
https://api.pushshift.io/reddit/search/comment/

## Contents
- [Import Libraries and Data](#Import-Libraries-and-Data)
- [Data Cleaning](#Data-Cleaning)
- [Save Data](#Save-Data)

## Import Libraries and Data

import regex as re
import pandas as pd

pd.set_option('display.max_columns', 40)
pd.set_option('display.max_rows', 100)

df_p = pd.read_csv("../datasets/politics_subreddit_latest_6500.csv")
df_c = pd.read_csv("../datasets/conservative_subreddit_latest_6500.csv")
df_l = pd.read_csv("../datasets/libertarian_subreddit_latest_6500.csv")

df_p

## Data Cleaning

### Drop Comments from the Comment Bot Labeled by Author Flair Text in r/politics

df_p.drop(df_p[df_p["author_flair_text"].str.contains("Bot", na = False, regex = True) == True].index, inplace = True)

### Check That No Comments from Comment the Bot Labeled by Author Flair Text in r/politics Remain

df_p[df_p["author_flair_text"].str.contains("Bot", na = False, regex = True) == True]

### Remove the Extra Column for Author Flair Text

df_p.drop(columns = "author_flair_text", inplace = True)
df_c.drop(columns = "author_flair_text", inplace = True)
df_l.drop(columns = "author_flair_text", inplace = True)

### Drop Null Values in Author Name Column from Deleted Posts

df_p = df_p.dropna()
df_c = df_c.dropna()
df_l = df_l.dropna()

print(df_p.isnull().sum())
print(df_c.isnull().sum())
print(df_l.isnull().sum())

### Check That No Removed Messages or Deleted Authors Remain

df_p[(df_p["body"] == "[removed]") | (df_p["author"] == "[deleted]")]

df_c[(df_c["body"] == "[removed]") | (df_c["author"] == "[deleted]")]

df_l[(df_l["body"] == "[removed]") | (df_l["author"] == "[deleted]")]

### Remove Formatting from Comment Bodies

def split_and_recombine_post(text):
    matches = re.findall("n*https*://[www.]*([A-Za-z0-9\.]*)/[\w+/\w+-.html]*|([A-Za-z0-9]+[\.]*[']*[A-Za-z]*)", text, re.MULTILINE)
    match_list = []
    for i in matches:
        if i[0]:
            match_list.append(i[0])
        if i[1]:
            match_list.append(i[1])
    recombined = " ".join(match_list)
    return recombined


for i in df_p.index:
    df_p.loc[i,"body"] = split_and_recombine_post(df_p.loc[i,"body"])
for i in df_c.index:
    df_c.loc[i,"body"] = split_and_recombine_post(df_c.loc[i,"body"])
for i in df_l.index:
    df_l.loc[i,"body"] = split_and_recombine_post(df_l.loc[i,"body"])

### Remove Any Posts That Had No Content in the Body

df_p.drop(df_p[df_p["body"] == ""].index, inplace = True)
df_c.drop(df_c[df_c["body"] == ""].index, inplace = True)
df_l.drop(df_l[df_l["body"] == ""].index, inplace = True)

## Save Data

df_p.to_csv("../datasets/politics_subreddit_latest_6500_cleaned.csv", index = False)
df_c.to_csv("../datasets/conservative_subreddit_latest_6500_cleaned.csv", index = False)
df_l.to_csv("../datasets/libertarian_subreddit_latest_6500_cleaned.csv", index = False)

## See Notebook 3 for Variable Creation
