==> code/Notebook-3-Create-Variables.ipynb <==
# Project 3: Web APIs & NLP
# Notebook 3: Create Variables

https://github.com/pushshift/api<br>
https://api.pushshift.io/reddit/search/comment/

## Contents
- [Import Libraries and Data](#Import-Libraries-and-Data)
- [Create and Save Variables](#Create-and-Save-Variables)

## Import Libraries and Data

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split

pd.set_option('display.max_columns', 40)
pd.set_option('display.max_rows', 100)

df_p = pd.read_csv("../datasets/politics_subreddit_latest_6500_cleaned.csv")
df_c = pd.read_csv("../datasets/conservative_subreddit_latest_6500_cleaned.csv")

## Create and Save Variables

### Add 1 for Liberal Slant (Comments from r/politics) and 0 for Conservative Slant (Comments from r/conservative) and Combine Datafames to Make a Single Training/Validation Dataframe

df_p["slant"] = 1
df_c["slant"] = 0

df_t = df_p.append(df_c, ignore_index = True)

df_t

df_t[df_t.isna().any(axis=1)]

### Define Variables, Establish Baseline, Train-Test-Split, and Save

X = df_t["body"]
y = df_t["slant"]

# there are slightly more liberal comments than conservative comments
# the baseline score is the number of liberal comments divided by the number of total comments
sum(y)/len(y)

X_train, X_test, y_train, y_test = train_test_split(X,
                                                    y,
                                                    test_size = 0.33,
                                                    stratify = y)

X_train.to_csv("../datasets/X_train", index = False)
X_test.to_csv("../datasets/X_test", index = False)
y_train.to_csv("../datasets/y_train", index = False)
y_test.to_csv("../datasets/y_test", index = False)

## See Notebook 4 for Modeling
